[impl] save position as float

+++ CloudBoard.ApiService/Dtos/NodeDto.cs
@@ -2,8 +2,8 @@ namespace CloudBoard.ApiService.Dtos;
 
 public class NodePositionDto
 {
-    public int X { get; set; }
-    public int Y { get; set; }
+    public float X { get; set; }
+    public float Y { get; set; }
 }
 
 public class NodeDto
